VMD script now accounts for both xyz and vmd files

+++ vmd/compil_vmd.sh
@@ -1,20 +1,21 @@
 for arg in "$@"; do
     directory=$(dirname $arg)
-    argnoxyz=${arg%.*}
-    file=$(basename $argnoxyz)
+    argnoext=${arg%.*}
+    file=$(basename $argnoext)
     
     tmp_script_name='tmp_vmd_script_for_'$file'.vmd'
-    cat ~/scripts/vmd/visualization_state_to_use.vmd > $tmp_script_name
-    
-    oldxyzfile=$(grep -h 'mol new .*xyz' $tmp_script_name | grep -o '.* type' | grep -o '.*xyz' | grep -o '/.*')
-    
-    currentdir=$(pwd) ; cd $directory ; newxyzfile=$(pwd) ; cd $currentdir
-    newxyzfile=$newxyzfile'/'$file'.xyz'
-    
-    sed -i "s|$oldxyzfile|$newxyzfile|g" $tmp_script_name
-    oldxyzfile=$(basename $oldxyzfile) ; newxyzfile=$(basename $newxyzfile)
-    sed -i "s|$oldxyzfile|$newxyzfile|g" $tmp_script_name
 
+    if [ -e $argnoext.vmd ]; then
+        cat $argnoext.vmd > $tmp_script_name
+    elif [ -e $argnoext.xyz ]; then
+        cat ~/scripts/vmd/visualization_state_to_use.vmd > $tmp_script_name
+        oldxyzfile=$(grep -h 'mol new .*xyz' $tmp_script_name | grep -o '.* type' | grep -o '.*xyz' | grep -o '/.*')
+        currentdir=$(pwd) ; cd $directory ; newxyzfile=$(pwd) ; cd $currentdir
+        newxyzfile=$newxyzfile'/'$file'.xyz'
+        sed -i "s|$oldxyzfile|$newxyzfile|g" $tmp_script_name
+        oldxyzfile=$(basename $oldxyzfile) ; newxyzfile=$(basename $newxyzfile)
+        sed -i "s|$oldxyzfile|$newxyzfile|g" $tmp_script_name
+    fi
     
     if [ -e ~/VMD_1_9_3/usr/local/lib/vmd/tachyon_LINUXAMD64 ]; then
         render='~/VMD_1_9_3/usr/local/lib/vmd/tachyon_LINUXAMD64'
